+++ erdjs/src/smartcontracts/codec.spec.ts
@@ -18,6 +18,7 @@ describe("test serialization", () => {
         definition.addField("current_ticket_number", PrimitiveType.U32);
         definition.addField("prize_pool", PrimitiveType.BigUInt);
 
+        // TODO: this is user defined...
         class Foo {
             ticket_price: BigInt = BigInt(0);
             tickets_left: number = 0;
@@ -48,27 +49,27 @@ describe("test serialization", () => {
     });
 });
 
-describe("test reader", () => {
-    it("should read big ints", async () => {
-        let data = serialized("00000008|8ac7230489e80000");
-        let reader = new BinaryReader(data);
-        let result = reader.readBigUInt();
-        assert.equal(result, BigInt("10000000000000000000"));
+// describe("test reader", () => {
+//     it("should read big ints", async () => {
+//         let data = serialized("00000008|8ac7230489e80000");
+//         let reader = new BinaryReader(data);
+//         let result = reader.readBigUInt();
+//         assert.equal(result, BigInt("10000000000000000000"));
 
-        data = serialized("0000000a|140ec80fa7ee88000000");
-        reader = new BinaryReader(data);
-        result = reader.readBigUInt();
-        assert.equal(result, BigInt("94720000000000000000000"));
+//         data = serialized("0000000a|140ec80fa7ee88000000");
+//         reader = new BinaryReader(data);
+//         result = reader.readBigUInt();
+//         assert.equal(result, BigInt("94720000000000000000000"));
 
-        // Now read both
-        data = serialized("[00000008|8ac7230489e80000] [0000000a|140ec80fa7ee88000000]");
-        reader = new BinaryReader(data);
-        assert.equal(reader.readBigUInt(), BigInt("10000000000000000000"));
-        assert.equal(reader.readBigUInt(), BigInt("94720000000000000000000"));
+//         // Now read both
+//         data = serialized("[00000008|8ac7230489e80000] [0000000a|140ec80fa7ee88000000]");
+//         reader = new BinaryReader(data);
+//         assert.equal(reader.readBigUInt(), BigInt("10000000000000000000"));
+//         assert.equal(reader.readBigUInt(), BigInt("94720000000000000000000"));
 
-        // TODO: Also read signed big ints.
-    });
-});
+//         // TODO: Also read signed big ints.
+//     });
+// });
 
 function serialized(prettyHex: string): Buffer {
     let uglyHex = prettyHex.replace(/[\|\s\[\]]/gi, "");
